requirements.txt
h5py==2.10.0
ipython==7.8.0
Keras==2.2.4
Keras-Applications==1.0.8
keras-lookahead==0.9.0
Keras-Preprocessing==1.1.0
keras-rectified-adam==0.17.0
matplotlib==3.3.2
nibabel==2.5.1
nipype==1.4.2
numpy==1.16.5
nvidia-ml-py3==7.352.0
nvidia-smi==0.1.3
opencv-python==4.2.0.32
pandas==1.1.4
scikit-learn==0.21.3
scipy==1.3.1
tensorboard==1.14.0
tensorflow-gpu==1.14.0

